Fixed bug in RLOG feature, typo in filament analysis
Former-commit-id: bb67b05c3e1df51eab25b9d40612c8e230c9417a

+++ src/simulation_manager.cpp
@@ -160,7 +160,7 @@ void SimulationManager::CheckRandomParams() {
    and returns N or 10^N, all depending on the random parameter generation type
    rtype, which can be R, RINT, or RLOG.
    *************************************/
-double SimulationManager::GetRandomParam(std::string rtype, int min, int max) {
+double SimulationManager::GetRandomParam(std::string rtype, double min, double max) {
   if (max == min) 
     error_exit("ERROR: Min and max value of parameter randomization sequence are equal.\n");
   if (rtype.compare("R") == 0)
